+++ suripu-core/src/test/java/com/hello/suripu/core/db/SleepSoundSettingsDynamoDBIT.java
@@ -1,6 +1,11 @@
 package com.hello.suripu.core.db;
 
-import com.amazonaws.services.dynamodbv2.model.CreateTableResult;
+import com.amazonaws.ClientConfiguration;
+import com.amazonaws.auth.BasicAWSCredentials;
+import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClient;
+import com.amazonaws.services.dynamodbv2.model.DeleteTableRequest;
+import com.amazonaws.services.dynamodbv2.model.ResourceInUseException;
+import com.amazonaws.services.dynamodbv2.model.ResourceNotFoundException;
 import com.google.common.base.Optional;
 import com.hello.suripu.core.db.sleep_sounds.SleepSoundSettingsDynamoDB;
 import com.hello.suripu.core.models.sleep_sounds.Duration;
@@ -8,7 +13,11 @@ import com.hello.suripu.core.models.sleep_sounds.SleepSoundSetting;
 import com.hello.suripu.core.models.sleep_sounds.Sound;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
@@ -17,17 +26,47 @@ import static org.hamcrest.core.Is.is;
  * Created by ksg 02/28/2017
  */
 
-public class SleepSoundSettingsDynamoDBIT extends DynamoDBIT<SleepSoundSettingsDynamoDB> {
+public class SleepSoundSettingsDynamoDBIT {
+    private final static Logger LOGGER = LoggerFactory.getLogger(SleepSoundSettingsDynamoDBIT.class);
 
-    @Override
-    protected CreateTableResult createTable() {
-        return dao.createTable(2L, 2L);
+    private AmazonDynamoDBClient amazonDynamoDBClient;
+    private SleepSoundSettingsDynamoDB dao;
+    private final String tableName = "test_table";
+
+
+    @Before
+    public void setUp(){
+        BasicAWSCredentials awsCredentials = new BasicAWSCredentials("FAKE_AWS_KEY", "FAKE_AWS_SECRET");
+        final ClientConfiguration clientConfiguration = new ClientConfiguration();
+        clientConfiguration.setMaxErrorRetry(0);
+        this.amazonDynamoDBClient = new AmazonDynamoDBClient(awsCredentials, clientConfiguration);
+        this.amazonDynamoDBClient.setEndpoint("http://localhost:7777");
+
+        cleanUp();
+
+        try {
+            LOGGER.debug("-------- Creating Table {} ---------", tableName);
+            SleepSoundSettingsDynamoDB.createTable(this.amazonDynamoDBClient, tableName);
+            LOGGER.debug("Created dynamoDB table");
+            this.dao = SleepSoundSettingsDynamoDB.create(this.amazonDynamoDBClient, tableName);
+        }catch (ResourceInUseException rie){
+            LOGGER.warn("Problem creating table");
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
     }
 
-    @Override
-    protected SleepSoundSettingsDynamoDB createDAO() {
-        return new SleepSoundSettingsDynamoDB(amazonDynamoDBClient, TABLE_NAME);
+    @After
+    public void cleanUp(){
+        final DeleteTableRequest deleteTableRequest = new DeleteTableRequest()
+                .withTableName(this.tableName);
+        try {
+            this.amazonDynamoDBClient.deleteTable(deleteTableRequest);
+        }catch (ResourceNotFoundException ex){
+            LOGGER.warn("Can not delete non existing table");
+        }
     }
+
 
     @Test
     public void testUpdateSetting() throws Exception {
@@ -38,7 +77,7 @@ public class SleepSoundSettingsDynamoDBIT extends DynamoDBIT<SleepSoundSettingsD
 
 
         final DateTime now = DateTime.now(DateTimeZone.UTC);
-        final Integer volumeFactor = 100;
+        final Integer volumeFactor = 90;
         final Duration duration = Duration.create(1L, "30 Minutes", 1800);
         final Sound sound = Sound.create(9L,
                 "https://s3.amazonaws.com/hello-audio/sleep-tones-preview/Brown_Noise.mp",
@@ -58,8 +97,9 @@ public class SleepSoundSettingsDynamoDBIT extends DynamoDBIT<SleepSoundSettingsD
         }
 
         final Duration duration2 = Duration.create(3L, "1 Hour", 3600);
+        final Integer volumeFactor2 = 57;
 
-        final SleepSoundSetting setting2 = SleepSoundSetting.create(senseId, accountId, now.plusMinutes(1), sound, duration2, volumeFactor);
+        final SleepSoundSetting setting2 = SleepSoundSetting.create(senseId, accountId, now.plusMinutes(1), sound, duration2, volumeFactor2);
         dao.update(setting2);
 
         final Optional<SleepSoundSetting> settingOptional2 = dao.get(senseId, accountId);
@@ -67,6 +107,7 @@ public class SleepSoundSettingsDynamoDBIT extends DynamoDBIT<SleepSoundSettingsD
         if (settingOptional2.isPresent()) {
             assertThat(settingOptional2.get().duration.name.equals(duration2.name), is(true));
             assertThat(settingOptional2.get().datetime.equals(now.plusMinutes(1)), is(true));
+            assertThat(settingOptional2.get().volumeScalingFactor, is(volumeFactor2));
         }
     }
